% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilers.R
\name{compile_bw_files}
\alias{compile_bw_files}
\title{Compile Body Weights Files}
\usage{
compile_bw_files(exp = NA, dir, opt = 1)
}
\arguments{
\item{exp}{a character string representing the name of the experiment}

\item{dir}{a character string representing the directory with the body weight files}

\item{opt}{an integer representing the option to process BW files. If Rosy Lane data (opt = 2)}
}
\value{
An Excel file with body weights
}
\description{
Compiles and formats body weights files.
}
