% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_milkcomp_files.R
\name{compile_milkcomp_files}
\alias{compile_milkcomp_files}
\alias{process_milkcomp_files}
\title{compile_milkcomp_files}
\usage{
process_milkcomp_files(file_list)
}
\arguments{
\item{Trial}{Name of the trial}

\item{Dir}{Directory with milk composition files in Excel format}
}
\value{
A table with the compiled milk composition files
}
\description{
Compile milk composition files
}
