% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilers.R
\name{compile_milkw_files}
\alias{compile_milkw_files}
\title{Compile Milk Weights Files}
\usage{
compile_milkw_files(exp = NA, dir)
}
\arguments{
\item{exp}{a character string representing the name of the experiment}

\item{dir}{a character string representing the directory with milk composition files in Excel format}
}
\value{
An Excel file with the compiled milk weights data
}
\description{
`compile_milkw_files()` compiles and formats milk weight files
}
