% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilers.R
\name{compile_VRfiles}
\alias{compile_VRfiles}
\title{Compile VR Files with Intakes}
\usage{
compile_VRfiles(dir, compfile, data)
}
\arguments{
\item{dir}{a character string representing the directory with compiled file}

\item{compfile}{a character string representing the name of the file with compiled data}

\item{data}{the data frame returned by `process_VRfiles()`}
}
\value{
An Excel file with compiled VR files
}
\description{
Compiles VR files with intakes from each day.
}
