% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{compiler}
\alias{compiler}
\title{Compile Files}
\usage{
compiler(exp = NA, dir, save_dir, type, compfile = NULL, vrdata = NULL)
}
\arguments{
\item{exp}{a character string representing the name of the experiment.}

\item{dir}{a character string representing the directory with files.}

\item{save_dir}{a character string representing the output directory.}

\item{type}{a character string indicating the type of data to process ('bw', 'mw', 'mcomp', 'vr').}

\item{compfile}{a character string indicating the name of the compiled file to use for merging (for 'vr' type).}

\item{vrdata}{a data frame with the process vr file}
}
\value{
A data frame with the compiled data.
}
\description{
Compiles and formats different types of files.
}
