% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_milkw_AFI.R
\name{format_milkw_AFI}
\alias{format_milkw_AFI}
\alias{process_AFI_files}
\title{format_milkw_AFI}
\usage{
process_AFI_files(file_list)
}
\arguments{
\item{Trial}{Name of the trial}

\item{Dir1}{Directory with files from AFI system (.txt)}

\item{Dir2}{Directory with csv files}
}
\value{
A table with the compiled milk weights from AFI system
}
\description{
Format milk weights from AFI.
}
