% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixers.R
\name{mixers}
\alias{mixers}
\title{Check Diet Intakes from RIC system}
\usage{
mixers(start_date, VRfolder, groups, bins = seq(1, 32))
}
\arguments{
\item{start_date}{A character string specifying the start date of the trial}

\item{VRfolder}{A character string with the path to the folder with VRfiles}

\item{groups}{A character string with the path to the file with groups IDs}

\item{bins}{Vector with the number of bins to be used. For example, 'seq(1,32)' or 'c(1,2,3,4,5,6, ..., 32)'}
}
\value{
A list of data sets with the intakes per diet per day, per week and per trial
}
\description{
Checking diets from VR files from RIC system with feed intakes
}
