% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_AFI_files}
\alias{process_AFI_files}
\title{process_AFI_files}
\usage{
process_AFI_files(Dir)
}
\arguments{
\item{Dir}{Directory with AFI files}
}
\value{
A table with all animals with milk weights
}
\description{
Process files from AFI system with milk weights
}
