% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_bw.R
\name{process_bw}
\alias{process_bw}
\title{Process Body Weights}
\arguments{
\item{file}{a character string representing the path to the Excel file containing body weights. The file should have columns:
'Visible_ID' (cow identifier), 'Date', 'BW_lbs' (body weight in pounds).}

\item{rfid_tbl}{a data frame with rfid and experiment information}

\item{start_date}{a character string representing the start date for the analysis (format 'mm/dd/yyyy')}

\item{end_date}{a character string representing the end date for the analysis (format 'mm/dd/yyyy')}
}
\value{
A list containing two data frames:
    \item{metabolic_bw}{Data frame containing metabolic body weights summary.}
    \item{delta_bw}{Data frame containing delta body weights.}
}
\description{
Process body weights files, and compute metabolic body weights and
    delta body weights using a linear model prediction for the entire period.
}
