% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_DC_milkw.R
\name{process_DC_milkw}
\alias{process_DC_milkw}
\title{Process and Format Dairy Comp Milk Weights}
\arguments{
\item{exp}{a character string representing the name of the experiment}

\item{file_path}{a character string representing the file path to Excel or CSV file downloaded from DC}

\item{late}{a numeric value representing the number of days late. By default 1}
}
\value{
A table with the milk weights from Dairy Comp system
}
\description{
Compiles and formats milk weight files from Dairy Comp system.
}
