% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_files}
\alias{process_files}
\title{process_files}
\usage{
process_files(Dir = NA, opt = 1)
}
\arguments{
\item{Dir}{A directory containing the file paths of the CSV, Excel, or no extension files to be processed.}

\item{opt}{Option to process Excel files from Rosy Lane (opt = 2)}
}
\value{
A data frame containing the combined data from all processed files.
}
\description{
Process csv, excel, or no extension files with specific format
   for milk composition and body weights files
}
