% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_intakes.R
\name{process_intakes}
\alias{process_intakes}
\title{Process Feed and Pellet Intakes}
\arguments{
\item{file1}{A string representing the path to the Excel file containing feed intake data. The file should have columns:
'Visible_ID' (cow identifier), 'Date' (date of intake), 'FedKg' (amount of feed in kilograms).}

\item{file2}{A string representing the path to the CSV file containing pellet intake data. The file should have columns:
'FarmName' (cow identifier), 'Date' (date of intake), 'PIntake_kg' (amount of pellets in kilograms).}

\item{rfid_tbl}{Data frame with rfid and experiment information}

\item{start_date}{A Date object representing the start date for the analysis.}

\item{end_date}{A Date object representing the end date for the analysis (currently not used in the function).}

\item{iDM}{A numeric vector representing the dry matter percentage for each week of the study. This should be as long as the number of weeks in the study.}

\item{pDM}{A numeric value representing the dry matter percentage of the pellet intake (default is 0.95).}
}
\value{
A data frame with the following columns:
    \item{Visible_ID}{Cow identifier.}
    \item{week}{Week number relative to the start date.}
    \item{n}{Number of intake records in that week.}
    \item{DMI}{Average dry matter intake (DMI) for the week.}
}
\description{
Process feed intakes and pellet inatkes. Creates a dataset
    with daily dry matter inatke for each animal across the study.
    The DMI includes individual feed and pellet intakes from RIC and GreenFeed system.
}
