% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIC2.R
\name{RIC2}
\alias{RIC2}
\title{Process VR files from RIC system}
\arguments{
\item{exp}{A character string specifying the experiment ID to use in file paths and processing.}

\item{VRfile}{VRfile with intakes}

\item{bins}{Vector with the number of bins to be used. For example, 'seq(1,32)' or 'c(1,2,3,4,5,6, ..., 32)'}

\item{output_dir}{The directory path where the output files should be saved.}

\item{output_file_name}{The name of the compiled Excel file. For example, "Intakes_XXX".}
}
\value{
A plot with the Fed in Kg per animal and an Excel file with the intakes.
}
\description{
Processes VR files from RIC system with feed intakes and
   save the results into an Excel file.
}
