% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_milke.R
\name{calculate_milke}
\alias{calculate_milke}
\title{Calculate Milk Energy}
\arguments{
\item{file1}{a character string representing the path to the Excel file containing milk weights}

\item{file2}{a character string representing the path to the Excel file containing milk composition}

\item{start_date}{a character string representing the start date for the analysis (format 'mm/dd/yyyy')}

\item{coeff1}{a numeric value representing the coefficient for fat (default is 9.29)}

\item{coeff2}{a numeric value representing the coefficient for protein (default is 5.85)}

\item{coeff3}{a numeric value representing the coefficient for lactose (default is 3.95)}
}
\value{
A data frame with daily milk energy
}
\description{
This function processes milk weights and composition data to calculate individual milk energy.
The calculation uses coefficients for fat, protein, and lactose to estimate milk energy.
}
