% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_bw_files.R
\name{compile_bw_files}
\alias{compile_bw_files}
\title{Compile Body Weights Files}
\usage{
compile_bw_files(Trial = NA, Dir = NA, opt = 1)
}
\arguments{
\item{Trial}{Name of the trial}

\item{Dir}{Directory where they are the body weight files}

\item{opt}{Option to process BW files. If Rosy Lane data (opt = 2)}
}
\value{
A table with body weights (AnimalID, Date, and Weights)
}
\description{
`compile_bw_files()` compiles and format body weights files.
}
