% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilers.R
\name{compile_milkcomp_files}
\alias{compile_milkcomp_files}
\title{Compile Milk Composition Files}
\usage{
compile_milkcomp_files(exp = NA, dir, output_dir)
}
\arguments{
\item{exp}{a character string representing the name of the experiment}

\item{dir}{a character string representing the directory with milk composition files in Excel format}

\item{output_dir}{a character string representing the output directory}
}
\value{
An Excel file with the compiled milk composition data
}
\description{
`compile_milkcomp_files()` compiles and formats milk composition files
}
