% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_milkcomp_files.R
\name{compile_milkcomp_files}
\alias{compile_milkcomp_files}
\title{Compile Milk Composition Files}
\usage{
compile_milkcomp_files(Trial = NA, Dir = NA)
}
\arguments{
\item{Trial}{Name of the trial}

\item{Dir}{Directory with milk composition files in Excel format}
}
\value{
A table with the compiled milk composition files
}
\description{
`compile_milkcomp_files()` compiles and formats milk composition files
}
