% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_milkw_AFI.R
\name{format_milkw_AFI}
\alias{format_milkw_AFI}
\title{format_milkw_AFI}
\arguments{
\item{Trial}{Name of the trial}

\item{Dir1}{Directory with files from AFI system (.txt)}

\item{Dir2}{Directory with csv files}
}
\value{
A table with the compiled milk weights from AFI system
}
\description{
`format_milkw_AFI()` compiles and formats milk weights files
    from AFI computer at the farm. First, it will compile cow files and
    then compile using the animal ID the milk weights with the previous ones.
}
\details{
This function uses two helper functions:
- `process_AFI_files`: Processes AFI files to extract and format milk weights.
}
