% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_AFI_files.R
\name{process_AFI_milkw}
\alias{process_AFI_milkw}
\title{Process and Format AFI Milk Weights}
\arguments{
\item{exp}{a character string representing the name of the experiment}

\item{dir1}{a character string representing the directory with files from AFI system (.txt)}

\item{dir2}{a character string representing the directory with CSV files}
}
\value{
A table with the compiled milk weights from AFI system
}
\description{
Compiles and formats milk weight files from AFI system.
    First, it compiles cow files, then compile using the ID the milk weights with the previous ones.
}
