% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_files}
\alias{process_files}
\title{Process Files}
\usage{
process_files(dir = NA, opt = 1)
}
\arguments{
\item{dir}{a character string representing the directory containing the file paths of the CSV, Excel, or no extension files to be processed}

\item{opt}{an integer representing the option to process Excel files from Rosy Lane (opt = 2)}
}
\value{
A data frame containing the combined data from all processed files.
}
\description{
Process CSV, Excel, or no extension files with specific format
   for milk composition and body weights files
}
